Add a rb-site option for adding additional entries for ALLOWED_HOSTS.
This adds a new `--allowed-host` setting to `rb-site`, which provides
another hostname or IP address for `settings.ALLOWED_HOSTS`. It's useful
for automated installs that need to configure a server that will
ultimately respond to multiple domain names.

Testing Done:
Manually tested using this with `rb-site install`, and in the new Docker
support.

Reviewed at https://reviews.reviewboard.org/r/11481/