Allow the common rb-site options to be used in subcommands.
When moving to `argparse`, the common set of options (`--noinput`,
`--debug`, etc.) were no longer able to be provided after the
subcommand. That is, `rb-site install --noinput` no longer worked.

This change addresses that by moving those options into a common
argparser and setting that as a "parent" for all others. This allows
them to all share the same set of options, allowing placement either
before or after the subcommand.

There's also a small tweak to one of the options, which had a wrong
description.

Testing Done:
Tested `rb-site install` and `rb-site upgrade` with common options
both before and after the subcommand, and verified they worked in
both cases.

Reviewed at https://reviews.reviewboard.org/r/11491/