Fix styling of legacy page_sidebar elements from extensions.
Review Board 4.0 introduced a new set of CSS component classes for
sidebars and page sidebars. In the process, it tried to provide
compatible styling for the legacy `#page_sidebar` element, which may be
used in some extensions (and in fact is used in Power Pack).

This didn't end up looking right, though. The page sidebar was lacking
some crucial positioning and padding, and this broke the Power Pack
Reports pages and likely any other extensions that might be using this.

This change adds some of the important legacy styles back. We're
hard-coding some positioning values, rather than using any existing
variables for these, because we want to ensure the layout is as close to
3.0's styles as possible, in order to ensure these pages won't break in
other subtle ways.

Testing Done:
Tested all the Power Pack 3 Reports pages and verified that layout was
correct once again.

Reviewed at https://reviews.reviewboard.org/r/11496/