Fix building the documentation for 4.0.
Documentation building had regressed, due to the need for finalizing a
setup after creating the database. Now, when building `docs.db`,
finalizing will take place.

This will still use the SCMTools defined in the fixtures, instead of
creating them when finalizing, to ensure the data shown in the
documentation is correct.