Keep inline Markdown code styles consistent between edit and render.
A pet peeve of mine for years has been that `inline code` would style
differently between CodeMirror (edit mode) and rendered Markdown. The
edit mode style was just a sort of metalic purple text. The rendered
style was red, with a faint background and border.

This change makes the styles consistent between the two, choosing the
rendered mode. There's now a 1-to-1 pixel mapping between the two, with
the exception that in edit mode, the backticks are visible.

Testing Done:
Tested in Chrome and Firefox. Swapped between edit and render modes, and
saw they were consistent.

Reviewed at https://reviews.reviewboard.org/r/11594/