Assume Subversion "nonexistent" destination revisions are deleted files.
Subversion has a couple different ways it can represent deleted files,
and they're not always consistent. It's also changed throughout
Subversion development, meaning that our deleted file logic worked fine
on some versions, and less so on others, depending.

We weren't handling the case where a file was listed as deleted using a
destination revision of `(nonexistent)`. We did handle this for the
source revision, and we handled `Index: ... (deleted)`, but not this
version.

This change implements this support, checking for this revision
identifier and flagging the file as deleted. This only affects
newly-uploaded diffs.

Testing Done:
Unit tests pass. Waiting on feedback from a user who reported the issue.

Reviewed at https://reviews.reviewboard.org/r/11602/