Add release notes for Review Board 4.0.
This covers all the features, improvements, and fixes that went into
Review Board 4.0, along with installation instructions, upgrade notes,
and comaptibility changes.

It's a big release with a lot to cover. The focus on the release notes
was to discuss the installation-related topics first, and to follow that
up with all the new feature changes, followed by bug fixes.

Testing Done:
Built the docs. Checked for build errors, link errors, and spelling
errors.

Reviewed at https://reviews.reviewboard.org/r/11606/