Fix CSS and stat precision issues with the Server Cache admin widget.
The Server Cache administration widget had some precision errors when
running on Python 3, due to differences in division. We now explicitly
opt into the `division` future, for Python 2.7, and explciitly format
the numbers in all cases. This gives us the same results as in Review
Board 3, with the addition of 2 digits of precision we now provide for
the percentages.

The lines in the widget are also now left-aligned. Due to a regression,
they were center-aligned in 4.0.0.

Testing Done:
Tested the widget on both Python 2.7 and 3.x, ensuring the appearance was
correct and the same between both.

Bugs closed: 4922

Reviewed at https://reviews.reviewboard.org/r/11614/