Clean up some of the extension hook unit tests and standardize setup.
This change simplifies some of the extension hook test setup code,
preparing for new hooks and unit tests. There's a new
`BaseExtensionHookTestCase` and `ExtensionHookTestCaseMixin` that
handles the main setup and teardown code for a test. This builds upon
the Djblets extension testing code, meaning we get all the benefits of
standardized and consistent extension management without ever having to
have tests specialize any of this.

There are some other small cleanups throughout for docstrings, or a little
bit of code or function names here and there. It's not a full revamp of
the code, just some tidying up.

Testing Done:
Unit tests passed.

Reviewed at https://reviews.reviewboard.org/r/11638/