Fix BaseRepositorySubForm to use new name for adding field prefixes.
In 1.11, Django renamed `Form.addprefix` to `Form.add_prefix`. When we
updated to it, we never made the change, and so the repository forms
were still calling the old function. This didn't affect any of the
built-in SCMTools, but for third-party tools that want to prefix keys
into the extra_data, it was failing.

Testing Done:
Used in conjunction with an external SCMTool.

Reviewed at https://reviews.reviewboard.org/r/11719/