Fix a regression in Subversion diff parsing.
A recent change intended to fix a deleted files indicator in Subversion
diffs regressed some parsing. We weren't checking whether the field in
question was `None`. This actually impacted a unit test, but it seems we
missed it (as CI is temporarily busted).

This change adds the necessary bullet-proofing, fixing the test.

Testing Done:
Applicable unit tests pass.

Tested with a customer.

Reviewed at https://reviews.reviewboard.org/r/11724/