Fix a crash when using Subvertpy 0.11.
Subvertpy 0.11 introduced some new logic in `client.propget()` that
normalizes a path to an absolute path. This in turn triggers an eventual
assertion failure in libsvn, which checks to make sure a full repository
URL wasn't passed in (somewhere). This was not a problem in Subvertpy
0.10 (which also didn't have this path normalization).

The problem applies only to `propget()`, and thankfully does not apply
to `proplist()`. To work around this crash, we're now calling
`proplist()` and filtering results from that. This method avoids the
path normalization that broke us, and provides the same results.

The Subvertpy bug is tracked at: https://github.com/jelmer/subvertpy/issues/35

Testing Done:
Unit tests pass on Python 2 and 3, against Subverpty 0.10 and 0.11.

Reviewed at https://reviews.reviewboard.org/r/11725/