Fix some documentation errors for SCMTool.parse_diff_revision().
`SCMTool.parse_diff_revision()` documented the strings as Unicode
strings, when in fact they're required to be byte strings. It also
didn't have sufficient documentation for the return type.

All the subclasses had correct documentation. The problem was just in
`SCMTool`.

This is a small change that fixes up this documentation to be correct.
It actually copy/pastes the corrected docs used in subclasses.

Testing Done:
Checked the documentation for errors.

Reviewed at https://reviews.reviewboard.org/r/11721/