Cap the version of the "packaging" module on Python 2.7 to < 21.0.
The "packaging" module, which provides some packaging functionality for
Python, dropped Python 2.7 support in v21.0. This is brought in through
third-party dependencies, so we need to cap it.

We now ensure that we install a version older than v21.0.

Testing Done:
Tested installing in a virtualenv. Got the latest version prior to v21.0.

Reviewed at https://reviews.reviewboard.org/r/11737/