Fix regressions in Download Diffs and Perforce unit tests.
We've had some broken tests recently, one of which was a regressions
from a merge, and the other a regression from potentially OpenSSL,
stunnel, or Perforce.

The first regression in the Download Diffs view was due to an old import
location for `TestService`. We now import from the correct place.

The second had to do with Perforce and stunnel. This might be
system-dependent, but locally, all the stunnel unit tests started
failing. Upon investigation, OpenSSL was complaining about the test
certificate we use for the stunnel proxy. Potentially, it was due to a
security issue present in the old certificate (length or cipher or
something). Generating a brand-new certificate appears to solve this
problem.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/11746/